Fix Preview Popup button not opening correctly
The Preview Popup button used to use the full value in the Subdomain
textbox to open, which results in an error if the user enters
"https://subdomain.onesignal.com" instead of just "subdomain". The
Preview Popup now auto-corrects the subdomain value in the same way the
web SDK auto-corrects the value.