Fix inconsistent checkbox state bug
Previously, users who checked "Send notification on post publish"
when the post was in a Draft or Pending Review state (or any non-
publish state), would later see the checkbox unchecked but a
notification would still go out. This was because the post meta-
data would have a flag to send out a notification, and once set,
this flag could not be unset.

This commit fixes the bug by using the Meta Box checkbox as the
source of truth for notifications sent from the WordPress editor.

If the checkbox is unchecked, then no notification is sent out (as
long as the notification was sent from WordPress's default editor).

For scheduled notifications, unchecking the checkbox and updating
the post will remove the meta flag.