Add proper OneSignal init WordPress action filter hook
The UI option to delay the initialization of OneSignal allows
customizing the web SDK initialization based on client-sided
parameters like the number of seconds the visitors has stayed
on the page or the number of visits.

But the UI option to delay the web SDK init does not allow
server-sided customization of the init option, in the way an
action/filter WordPress hook can.

So this commit adds a WordPress action/filter hook for the
initialization of the web SDK.