Clarify when a user has opted to delay initializing the SDK
The UI option "Use my own custom SDK initialization script" prints the plugin's
configuration to the page as window.__oneSignalInitOptions, but prevents the
plugin from actually initializing the web SDK.

It can be hard to know when a user has accidentally or intentionally turned this
on, so if a user has opted to delay initialization the SDK, a comment is printed
on the page instead. This should help with debugging sites.