Remove Preview Popup button
It doesn't seem like it's been used that much (since the link it opened
was broken for the last few months) and we only got one mention about
this. The section header image can be matched with the textbox fields
instead and while it's a little less convenient, it's still usable to
know what values you're changing.