Hide Google Project Number from configuration
Users who try to add an Android or iOS project to their OneSignal app
get confused and try to add a Google Project Number to their site, which
results in sending errors since  we expect the site to use our default
Project Number. This commit adds a special query flag
'reveal_project_number=true' that has to be added to the page in order
to reveal the Project Number setting.