Add deprecation notice
Motivation: shows a universal admin-level notice (not just in OS dashboard page) that we are deprecating OS config in this plugin and are moving it to OneSignal.com

If dismissed, we will re-show the notice after 7 days. We will continue re-showing the notice until December 31st 2024 to give a bit of a buffer.