feat: uncheck the "Send notification" checkbox when a post is published
We accomplish this by polling for the status of the post every second for up to ten seconds. When a post has completed posting/saving, then we will uncheck the "Send notification" checkbox to ensure that the state is unaffected prior to composing & sending the notification request to OneSignal. This only applies to publishing a post for the first time.

TODO: need to investigate and implement similar logic for updating a post