feat: uncheck the "Send notification" checkbox when a post is republished
Prior to this change we would run into an edge case where a notification was saved as already "sent" (via the `_onesignal_notification_sent` global). This caused notifications to not be sent on alternating instances of updating a post. This commit aims to address that by adding an AJAX endpoint to reset notification status *after* successfully sending a notification.

Additionally, this commit removes duplicate script enqueueing introduced earlier as it was causing duplicate requests to be made to the admin-ajax endpoints. It also adds some improvements to polling, such as error handling, to help with debugging