fix: move upgrade logic block to migrated apps
Previously, 7c82c31 added the logic to default the `notification_on_page` to on if the app was *not* migrated and *not* a new install. For the case of upgrades between v3 versions, this was incorrect.

Now, the default will only be set if the app is not a fresh install (preliminary settings already exist).