feat: cancel previous scheduled notifications when posts are updated
Previously, if a user scheduled a post with the "sent notification" checkbox enabled it would create a scheduled notification. If the user then modified the scheduled post and pressed Save, it would create a new scheduled notification but the previous one would still remain active. This could result in multiple notifications being sent, likely one with an incorrect `contents` heading.

Now, the onesignal_schedule_notification logic will store the original notification ID and cancel it if any updates are saved to the same post.